aiagentapi/requirements.txt
# 共通
python-dotenv
requests
# FastAPI
fastapi
uvicorn
# langchain & tool
langchain
langchain-openai
google-genai
google-generativeai
beautifulsoup4
langgraph
langchainhub
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
html2text
langchain-google-genai
pydantic-settings
openai
pandas
# itsdangerous
# adk
google-adk
google-genai
httpx
# LangGraph
langchain-mcp-adapters
langgraph
langgraph-sdk
langchain-openai